% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_order_to_dendrogram.R
\name{convert_order_to_dendrogram}
\alias{convert_order_to_dendrogram}
\title{Convert Order to Dendrogram}
\usage{
convert_order_to_dendrogram(sampleorder)
}
\description{
\code{convert_order_to_dendrogram} is a helper function used in the \code{heatmap_generator} function that creates a flat dendrogram based on an ordered list of integers.
}
