% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_generator.R
\name{heatmap_generator}
\alias{heatmap_generator}
\title{AGED Heatmap Generator}
\usage{
heatmap_generator(
  aged_results,
  data,
  samp_info,
  batches = names(samp_info),
  clear_low_variance = FALSE,
  transformation_type = "",
  blind = TRUE,
  pearson = FALSE,
  specific_order = NULL,
  legend = TRUE,
  dendrogram = "none",
  trace = "none",
  scale = "row",
  cexRow = 0.5,
  key = FALSE,
  lhei = c(1, 3),
  lwid = c(2, 3),
  legend_size = 0.75,
  legend_space = 1,
  ...
)
}
\arguments{
\item{aged_results}{The results of a successful call to \code{AGED}}

\item{data}{The original data that was plugged into the initial call of \code{AGED}}

\item{samp_info}{A matrix-like object containing sampling information regarding the samples used in the original AGED call. A proper samp_info object will have \code{n} rows corresponding to the \code{n} samples, and columns comprised of information pertaining to the samples such as extraction date, other characteristics, etc.}

\item{batches}{A string vector that details the tracks of characteristics to outline on the heatmap. Each element of the string vector must correspond to a column name of samp_info.}

\item{clear_low_variance}{A boolean variable that determines whether rows with var < 1 are removed or not. This is done before any type of transformation is performed.}

\item{transformation_type}{A string variable that determines whether or not a log or VST transformation should be done on the original data. If this argument is used, it should be "vst" or "log" only. If no transformation is to be performed, the default value or any string other than "vst" or "log" can be used.}

\item{blind}{If a VST is to be done, this boolean value determines whether it is blind or not.}

\item{pearson}{A boolean value indicating whether or not pearson (row) clustering should be performed.}

\item{specific_order}{A vector of strings representing how the sample tracks (columns) should be sorted. Each element of this vector should represent a column name in \code{batches}. The sample tracks will be nested sorted by the first element first, then the second element, etc.}

\item{legend}{A boolean value indicating whether or not a legend should be plotted alongside the heatmap. WARNING: Adding a row dendrogram will interfere with the legend if plotted concurrently.}

\item{dendrogram}{A character string indicating whether to draw "none", "row", "column" or "both" dendrograms on the heatmap. Defaults to "column". This argument should only be one of these strings. WARNING: Adding a row dendrogram will interfere with the legend if plotted concurrently.}

\item{trace}{A character string indicating whether a solid trace line should be drawn across the "row"s, down the "column"s, "both", or "none". The distance of the line from the center of each color-cell is proportional to the size of the measurement. Defaults to "none". This argument should only be one of these strings.}

\item{scale}{A character string indicating if the values should be centered and scaled in the "row" direction, the "column" direction, or "none". The default is "none". This argument should only be one of these strings.}

\item{cexRow}{A positive number, used as 'cex.axis' for column axis labeling.}

\item{key}{A boolean value indicating whether or not a color key should be drawn.}

\item{lhei}{A numerical vector indicating the relative row heights of the rows of the plot.}

\item{lwid}{A numerical vector indicating the relative column widths of the rows of the plot.}

\item{legend_size}{A numerical value indicating the size of the legend. The default value is 0.75. Increasing this value will increase the text and block size of the legend.}

\item{legend_space}{A numerical value indicating how far spaced apart different elements in the legend should be. The default value is 1. Increasing this value will increase the space between different elements in the legend.}

\item{...}{Optional arguments that can be passed to \code{heatmap.3}. For a full list of arguments, check \href{https://github.com/obigriffith/biostar-tutorials/blob/master/Heatmaps/heatmap.3.R}{here}.}
}
\value{
Plots the requested heatmap
}
\description{
\code{heatmap_generator} will pull metagene information from \code{AGED} results to create a heatmap. By default space for four components will be allocated in the plot. At the top left will be the color key, the top right will have the column dendrogram, the bottom left will have the row dendrogram, and the bottom right will have the image plot.
}
