% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGSE75214_IBD.R
\name{parseGSE75214}
\alias{parseGSE75214}
\title{Parse IBD Dataset}
\usage{
parseGSE75214()
}
\description{
\code{parseGSE75214} parses the IBD data fileand allows it to be ready for use by \code{AGED}.
}
