% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_and_transform_data.R
\name{verify_and_transform_data}
\alias{verify_and_transform_data}
\title{Transform Data}
\usage{
verify_and_transform_data(
  data,
  clear_low_variance = FALSE,
  transformation_type = "",
  blind = TRUE
)
}
\arguments{
\item{data}{The original data that was plugged into the initial call of \code{AGED}}

\item{clear_low_variance}{A boolean variable that determines whether rows with var < 1 are removed or not. This is done before any type of transformation is performed.}

\item{transformation_type}{A string variable that determines whether or not a log or VST transformation should be done on the original data. If this argument is used, it should be "vst" or "log" only. If no transformation is to be performed, the default value or any string other than "vst" or "log" can be used.}

\item{blind}{If a VST is to be done, this boolean value determines whether it is blind or not.}
}
\value{
Returns the transformed data.
}
\description{
\code{verify_and_transform_data} is a helper function used in multiple core functions to check input, clear low variance and apply log or variance stabilizing transformation.
}
