% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aged.R
\name{aged}
\alias{aged}
\title{Automatic Gene Expression Deconvolution Using FaStaNMF}
\usage{
aged(
  data,
  rank,
  n = 25,
  nrun = 200,
  nmf_seed = 123456,
  mvg = 1000,
  clv = 0,
  transformation = 0,
  blind = TRUE,
  ...
)
}
\arguments{
\item{data}{Gene expression target data, a matrix-like object. The rows should represent genes, and each row must have a unique row name. Each column should represent a different sample.}

\item{rank}{The factorization rank (number of factors) to be used during NMF. This function argument should be a positive integer value.}

\item{n}{The number of barcode genes desired per metagene}

\item{nrun}{The desired number of NMF runs to be run on the initially reduced dataset.}

\item{nmf_seed}{The desired seed to be used for NMF on the initially reduced dataset.}

\item{mvg}{A numerical argument determining how many of the most variable genes to look at during the first steps of FaStaNMF.}

\item{clv}{A numerical value \code{x} that reduces the dataset by removing genes with variance < \code{x} across all samples. Our recommended value is to set this parameter to 1 if genes expression low variance across samples is desired. These genes will not be considered at all for the deconvolution. This is done before any type of transformation or other reduction is performed.}

\item{transformation}{A numerical value that determines whether or not a log or VST transformation should be done on the original dataset. A value of 0 indicates no transformation, a value of 1 indicates a log transformation using \link[base]{log1p}, a value of 2 indicates a VST transformation using \link[DESeq2]{varianceStabilizingTransformation} If this argument is used, it should be "0", "1" or "2" only. Any other value will assume no transformation. For FaStaNMF, untransformed data should be log-transformed or VST-transformed.}

\item{blind}{If a VST is to be done using the \code{transformation} parameter, this boolean value determines whether it is blind or not.}

\item{...}{Other arguments to be passed to FaStaNMF.}
}
\value{
A list containing barcode genes for each metagene, and the raw W and H matrix returned by FaStaNMF.
}
\description{
\code{AGED} will perform gene expression deconvolution on a matrix or data frame of gene expression data using FaStaNMF.
}
